submit order logic added
When a user submits an order to the table, it changes the table to have an order, changes the status to 'active' and will update the button to be a different color.

The tables that are active no longer show up in the 'new table' page and will now show up on the 'my tables' page!